# drakkar/uiserver/openapi.yaml
...
openapi: 3.1.0
info:
  title: Drakkar Backend API
  version: "1.2"
  description: >-
    The JSON surface every Drakkar backend (Python reference and Go port)
    serves identically under /api/v1, plus the unprefixed Kubernetes probes.
    One optional bearer token protects every /api/* route: configure
    ui.auth_token and send `Authorization: Bearer <t>` or `?token=<t>`;
    with no token configured, auth is disabled (private-contour default).


    Not modeled by OpenAPI but part of the contract: `GET /ws` upgrades to a
    WebSocket streaming recorder events — accept, then token check (close
    4401) and origin check (close 4403); at most 100 events per drain pass,
    20 ms idle backoff, one JSON recorder-event object per text frame.


    Legacy unprefixed `/api/...` aliases exist for backends' built-in pages
    and are deliberately outside this document; the SPA only calls /api/v1.
servers:
  - url: /
    description: Same-origin — the backend serves both the SPA and the API.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/debug/runtime/units:
    get:
      tags: [debug-tools]
      operationId: getRuntimeUnits
      summary: Concurrency-unit census
      description: >-
        Live concurrency units grouped by (name, suspension point) —
        asyncio tasks on the Python backend, goroutines on the Go backend.
        Runs on the measured runtime itself; 503 means the dispatch timed
        out because the runtime is not serving work (stalled), which is
        itself a diagnostic answer.
      responses:
        '200':
          description: Census grouped by suspension point, largest first.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RuntimeUnitCensus'}
        '503':
          description: Census dispatch timed out — the runtime is stalled.
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/debug/kafka/topics:
    get:
      tags: [kafka-read]
      operationId: getKafkaTopics
      summary: Readable Kafka topic aliases
      description: >-
        The configured read aliases — `source`, `dlq`, and each Kafka sink
        by instance name. Raw topic names and broker addresses are never
        exposed; the alias is the entire addressing surface. 403
        {"error":...} when ui.kafka_read_enabled=false.
      responses:
        '200':
          description: The alias list.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KafkaReadTopics'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/debug/kafka/{alias}/message:
    get:
      tags: [kafka-read]
      operationId: getKafkaMessage
      summary: One Kafka message by coordinates
      description: >-
        Fetches exactly one message by (partition, offset) from the aliased
        topic via an ad-hoc assign()-only consumer — no consumer group is
        joined, no offsets move. 404 for an unknown alias or partition, an
        offset outside the current watermarks, or a compacted/deleted
        record; 502 when the brokers do not answer within the deadline.
      parameters:
        - name: alias
          in: path
          required: true
          schema: {type: string}
        - name: partition
          in: query
          required: true
          schema: {type: integer, minimum: 0}
        - name: offset
          in: query
          required: true
          schema: {type: integer, minimum: 0}
      responses:
        '200':
          description: The message with full record metadata.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KafkaReadMessage'}
        '404':
          description: Unknown alias/partition, offset out of range, or record gone.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '502':
          description: Brokers unreachable or fetch deadline elapsed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/debug/kafka/{alias}/messages:
    get:
      tags: [kafka-read]
      operationId: getKafkaMessages
      summary: Stream a time window of Kafka messages (NDJSON)
      description: >-
        application/x-ndjson — one KafkaReadMessage JSON object per line,
        oldest-first within each partition, best effort across partitions.
        Starts at the first offsets at/after from_ts (broker-side
        offsets-for-times) and ends at to_ts, after limit messages (server
        cap 10000), or at the end-of-topic snapshot taken when the request
        started — whichever comes first. A mid-stream broker failure
        terminates the stream with a final {"error": "..."} line (the 200
        header is already committed). Same ad-hoc no-group consumer
        contract as the single-message endpoint.
      parameters:
        - name: alias
          in: path
          required: true
          schema: {type: string}
        - name: from_ts
          in: query
          required: true
          schema: {type: integer, minimum: 0}
        - name: to_ts
          in: query
          schema: {type: integer, minimum: 0}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 10000}
        - name: partition
          in: query
          schema: {type: integer, minimum: 0}
      responses:
        '200':
          description: NDJSON stream of KafkaReadMessage lines.
          content:
            application/x-ndjson:
              schema: {$ref: '#/components/schemas/KafkaReadMessage'}
        '404':
          description: Unknown alias or partition.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '502':
          description: Brokers unreachable while resolving the window.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Detail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Only enforced when ui.auth_token is configured; otherwise every
        route is open (private-contour default). The token may also ride as
        a ?token= query parameter (downloads, WebSocket).
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (only when auth is configured).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Detail'}
    ValidationError:
      description: Invalid parameters or body.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ValidationDetail'}
  schemas:
...
    Detail:
      description: Standard error envelope for 401/404/500/503.
      type: object
      properties:
        detail: {type: string}
      required: [detail]
    ValidationDetail:
      description: Pinned-minimal 422 envelope — consumers read only msg.
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items: {}
              msg: {type: string}
            required: [loc, msg]
      required: [detail]
...
    KafkaReadTopics:
      type: object
      properties:
        topics:
          type: array
          items:
            type: object
            properties:
              alias: {type: string}
              kind: {type: string, enum: [source, dlq, sink]}
            required: [alias, kind]
      required: [topics]
    KafkaReadHeader:
      type: object
      properties:
        key: {type: string}
        value: {type: [string, 'null']}
        value_encoding: {type: [string, 'null'], enum: [utf-8, base64, null]}
      required: [key, value, value_encoding]
    KafkaReadMessage:
      type: object
      description: >-
        One Kafka record. Byte fields (payload, key, header values) carry
        utf-8 text when the bytes decode, otherwise base64, with the
        encoding flagged alongside. alias echoes the configured read
        alias; the raw topic name is deliberately absent from the wire.
      properties:
        alias: {type: string}
        partition: {type: integer}
        offset: {type: integer}
        timestamp_ms: {type: [integer, 'null']}
        key: {type: [string, 'null']}
        key_encoding: {type: [string, 'null'], enum: [utf-8, base64, null]}
        payload: {type: string}
        payload_encoding: {type: string, enum: [utf-8, base64]}
        payload_size_bytes: {type: integer}
        headers:
          type: array
          items: {$ref: '#/components/schemas/KafkaReadHeader'}
      required: [alias, partition, offset, timestamp_ms, key, key_encoding,
                 payload, payload_encoding, payload_size_bytes, headers]
...
    RuntimeUnitGroup:
      description: Units sharing one (name, suspension point) pair.
      type: object
      properties:
        name: {type: string, description: Coroutine / function qualname.}
        location: {type: string, description: file:line of the suspension point ("" when unknown).}
        count: {type: integer}
        example: {type: string, description: Name of one representative unit.}
      required: [name, location, count, example]
    RuntimeUnitCensus:
      type: object
      properties:
        unit_label: {type: string}
        total: {type: integer}
        units:
          type: array
          items: {$ref: '#/components/schemas/RuntimeUnitGroup'}
      required: [unit_label, total, units]
...